requirements.txt
absl-py==2.1.0
astor==0.8.1
astunparse==1.6.3
backports.zoneinfo==0.2.1
beautifulsoup4==4.13.3
certifi==2025.1.31
charset-normalizer==3.4.1
cloudpickle==3.1.1
docopt==0.6.2
gast==0.3.3
google-pasta==0.2.0
grpcio==1.70.0
h5py==2.10.0
idna==3.10
importlib_metadata==8.5.0
joblib==1.4.2
Js2Py==0.74
keras==2.15.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
Markdown==3.7
MarkupSafe==2.1.5
mtcnn==0.1.1
numpy==1.19.5
nvidia-cublas==11.4.1.1026
nvidia-cuda-cupti==11.2.135
nvidia-cuda-nvcc==11.2.142
nvidia-cuda-nvrtc==11.1.105
nvidia-cuda-runtime==11.2.146
nvidia-cudnn==8.1.1.33
nvidia-cufft==10.4.0.135
nvidia-curand==10.2.3.135
nvidia-cusolver==11.1.0.135
nvidia-cusparse==11.4.0.135
nvidia-dali-cuda110==0.31.0
nvidia-dali-nvtf-plugin==0.31.0+nv21.3
nvidia-horovod==0.21.3+nv21.3
nvidia-nccl==2.8.4
nvidia-pyindex==1.0.9
nvidia-tensorboard==1.15.0+nv21.3
nvidia-tensorflow==1.15.5+nv21.3
nvidia-tensorrt==7.2.2.3
opencv-python==4.7.0.72
opt_einsum==3.4.0
packaging==24.2
pandas==1.0.0
pillow==10.4.0
protobuf==3.20.1
psutil==7.0.0
pyjsparser==2.7.1
PyPrind==2.11.3
pySmartDL==1.3.4
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.2
requests==2.32.3
scikit-learn==0.23.0
scipy==1.10.1
six==1.17.0
soupsieve==2.6
tensorboard @ https://pypi.ngc.nvidia.com/tensorboard/tensorboard-1.15.0-py2.py3-none-any.whl#sha256=413865fcdf03e693de43372be1fe64b37f0982d792cbd80bbec2bd06572413ac
tensorflow-estimator==1.15.1
termcolor==2.4.0
threadpoolctl==3.5.0
torch==1.8.2+cu111
torchaudio==0.8.2
torchvision==0.9.2+cu111
tqdm==4.66.1
typing_extensions==4.12.2
tzlocal==5.2
urllib3==2.2.3
webencodings==0.5.1
Werkzeug==3.0.6
wrapt==1.17.2
zipp==3.20.2
